Updating tests to match new duration/renewbefore format
