+++ test/util/generate/issuer.go
@@ -44,8 +44,6 @@ func Issuer(cfg IssuerConfig) *v1alpha1.Issuer {
 		},
 		Spec: v1alpha1.IssuerSpec{
 			IssuerConfig: v1alpha1.IssuerConfig{
-				Duration:    cfg.Duration,
-				RenewBefore: cfg.RenewBefore,
 				ACME: &v1alpha1.ACMEIssuer{
 					SkipTLSVerify: cfg.ACMESkipTLSVerify,
 					Server:        cfg.ACMEServer,
